requirements.dev.txt
imagehash>=3.4
ipdb
Pillow>=4.0.0
pycparser>=2.17.0
pytest>=3.0.5
pytest-instafail==0.3.0
scipy>=0.18.0
sphinx
sphinx_rtd_theme
numpydoc
